refactor: #674 mui-chat をコンポーネント/hooks 分割する
UI とセッション/送信ロジックを分離し、page 向け export は維持する。
選択肢抽出のユニットテストを追加する。
