+++ frontend/components/mui-chat/components/TypingIndicator.tsx
@@ -1,0 +1,34 @@
+'use client'
+
+import { Box, CircularProgress, Typography } from '@mui/material'
+
+/** AI応答待ちのタイピングインジケーター */
+export function TypingIndicator() {
+  return (
+    <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+      <CircularProgress size={16} />
+      <Typography variant="body2" color="text.secondary">
+        AIが考えています
+      </Typography>
+      <Box sx={{ display: 'flex', gap: 0.5 }}>
+        {[0, 0.16, 0.32].map((delay, i) => (
+          <Box
+            key={i}
+            sx={{
+              width: 6,
+              height: 6,
+              borderRadius: '50%',
+              bgcolor: 'text.secondary',
+              animation: 'bounce 1.4s infinite ease-in-out',
+              animationDelay: `${delay}s`,
+              '@keyframes bounce': {
+                '0%, 80%, 100%': { transform: 'scale(0)' },
+                '40%': { transform: 'scale(1)' },
+              },
+            }}
+          />
+        ))}
+      </Box>
+    </Box>
+  )
+}
